+++ src/Symfony/Component/Uid/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+7.1
+---
+
+ * Add `UuidV1::toV6()`, `UuidV1::toV7()` and `UuidV6::toV7()`
+
 6.2
 ---